Added MOS and BAM file supporting, from Infinity Engine.
git-svn-id: https://cvet.by:8443/svn/FOnlineSource@196 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: 10c5e0cabce8c50247522c7b1ff40b4658259e5d

+++ Source/FileManager.cpp
@@ -328,7 +328,7 @@ bool FileManager::GetLine(char* str, uint len)
 	return true;
 }
 
-bool FileManager::CopyMem(void* ptr, size_t size)
+bool FileManager::CopyMem(void* ptr, uint size)
 {
 	if(!size) return false;
 	if(curPos+size>fileSize) return false;
